Rename utilities -> utils + formatting.
